workaround a bug in python imaging 1.1.7 (AttributeError: 'NoneType' object has no attribute 'bands')
The problem is that PIL should load the image before trying to access
the number of bands, but doesn't. This is known upstream (there is a
debian bug against PIL), to work around it you just need to do the
load() explicitly before trying to access the number of bands.