link with -lm if possible
this fixes the "undefined reference to `log10' issue on mageia linux.
see https://github.com/pediapress/timelib/issues/3 for the same issue
in timelib or the recent thread on the mailing list.

mageia linux links cpython modules with -Wl,--no-undefined, which
makes the linker believe log10 is not defined. That's right, but when
the cpython module is dlopen'ed the symbol would be there cause python
is linked against the math lib.

thanks cescp for reporting!